pyproject.toml
[project]
name = "chaster-brain"
version = "0.1.0"
description = "External backend intelligence layer for Chaster."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "pyjwt[crypto]>=2.9.0",
  "pypdf>=5.4.0",
  "langgraph>=0.2.15",
  "llama-index>=0.11.0",
  "bleach>=6.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
